pyproject.toml
[project]
name = "mkk-moon"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.0",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.38.0",
]
